fix: fix eslint issue and move file into correct folder
