+++ .eslintrc.json
@@ -20,6 +20,7 @@
     "simple-import-sort/imports": "error",
     "simple-import-sort/exports": "error",
     "prettier/prettier": "error",
+    "no-nested-ternary": "warn",
     "react/react-in-jsx-scope": "off",
     "jsx-a11y/anchor-is-valid": "off",
     "react/require-default-props": "off",
