--- .yarnrc 2.yml
@@ -1,3 +1,0 @@
-nodeLinker: node-modules
-
-yarnPath: .yarn/releases/yarn-4.5.1.cjs
