+++ src/domain/props/TypingTextProps.tsx
@@ -1,0 +1,7 @@
+// TypingTextProps is a type that defines the props for the TypingText component
+export default interface TypingTextProps {
+  text: string;
+  speed?: number;
+  endSentencePause?: number;
+  onTypingComplete?: () => void;
+}
